+++ src/lib/gallery.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { NgxGalleryService } from './services/ngx-gallery.service';
+import { GalleryService } from './services/gallery.service';
 
-describe('NgxGalleryService', () => {
-  let service: NgxGalleryService;
+describe('GalleryService', () => {
+  let service: GalleryService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(NgxGalleryService);
+    service = TestBed.inject(GalleryService);
   });
 
   it('should be created', () => {