+++ src/lib/interface/index.ts
@@ -1,6 +1,6 @@
-export type { NgxGalleryOptions } from './ngxGalleryOptions.interface'
+export type { GalleryOptions, NgxGalleryOptions } from './gallery-options.interface'
 export type { DialogShareOptions } from './dialogShare.interface'
-export type { NgxGalleryImage } from './ngxGalleryImage.interface'
+export type { GalleryImage, NgxGalleryImage } from './gallery-image.interface'
 
 export type { CustomStyles, Border } from './customStyles.interface'
 export type { LightboxImage } from './image-object.interface'